Moved business logic for receiving delivery options to OrderController. Removed unnecessary classes
