--- src/main/java/com/hll/hyperlightlogistics/config/KafkaTopicConfig.java
@@ -1,36 +1,0 @@
-/*
- * This file is part of HyperLightLogistics-Java.
- *
- * HyperLightLogistics-Java is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * HyperLightLogistics-Java is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with HyperLightLogistics-Java.  If not, see <https://www.gnu.org/licenses/>.
- *
- * Copyright (C) 2024 Rauf Agaguliev
- */
-
-package com.hll.hyperlightlogistics.config;
-
-import org.apache.kafka.clients.admin.NewTopic;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
-
-@Configuration
-public class KafkaTopicConfig {
-
-//    @Bean
-//    public NewTopic deliveryOptionsResponseTopic() {
-//        return TopicBuilder.name("delivery-options-request-topic")
-//                .build();
-//    }
-
-}
