--- src/main/java/com/hll/hyperlightlogistics/controller/DeliveryController.java
@@ -1,47 +1,0 @@
-/*
- * This file is part of HyperLightLogistics-Java.
- *
- * HyperLightLogistics-Java is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * HyperLightLogistics-Java is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with HyperLightLogistics-Java.  If not, see <https://www.gnu.org/licenses/>.
- *
- * Copyright (C) 2024 Rauf Agaguliev
- */
-
-package com.hll.hyperlightlogistics.controller;
-
-import com.hll.hyperlightlogistics.dto.DeliveryRequestDTO;
-import com.hll.hyperlightlogistics.dto.ProductDeliveryOptionDTO;
-import com.hll.hyperlightlogistics.service.DeliveryService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/v1/delivery")
-public class DeliveryController {
-
-    @Autowired
-    private DeliveryService deliveryService;
-
-    @PostMapping("/options")
-    public ResponseEntity<List<ProductDeliveryOptionDTO>> getDeliveryOptions(@RequestBody DeliveryRequestDTO request) {
-        List<ProductDeliveryOptionDTO> deliveryOptions = deliveryService.calculateDeliveryOptions(request);
-        return ResponseEntity.ok(deliveryOptions);
-    }
-
-}
