--- src/main/java/com/hll/hyperlightlogistics/kafka/KafkaConsumer.java
@@ -1,33 +1,0 @@
-/*
- * This file is part of HyperLightLogistics-Java.
- *
- * HyperLightLogistics-Java is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * HyperLightLogistics-Java is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with HyperLightLogistics-Java.  If not, see <https://www.gnu.org/licenses/>.
- *
- * Copyright (C) 2024 Rauf Agaguliev
- */
-
-package com.hll.hyperlightlogistics.kafka;
-
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class KafkaConsumer {
-
-    @KafkaListener(topics = "delivery-options-response-topic", groupId = "delivery_group")
-    public void listenDeliveryOptions(String message) {
-        System.out.println("Received delivery options: " + message);
-    }
-
-}
